Bug: Placing float type bids

+++ src/Auction/AuctionItemBrowse.js
@@ -135,7 +135,7 @@ const ItemBrowse = forwardRef(({ onBidSubmit, sortType, items, bidAmounts, setBi
         const bidAmount = bidAmounts[itemId];
         const userId = localStorage.getItem("userId");
         const minBidAmount = document.getElementById(`bid-${itemId}`).min;
-        if (bidAmount < parseInt(minBidAmount) + 0.5) {
+        if (bidAmount < parseFloat(minBidAmount) + 0.5) {
             setMessageAlert(`Please place a bid amount of at least \u00A3${parseFloat(minBidAmount) + 0.5}`);
             setBidAmounts({
                 ...bidAmounts,
